+++ httpie/core.py
@@ -204,4 +204,8 @@ def main(args=sys.argv[1:], env=Environment(), error=None):
         if downloader and not downloader.finished:
             downloader.failed()
 
+        if (not isinstance(args, list) and args.output_file and
+                args.output_file_specified):
+            args.output_file.close()
+
     return exit_status