+++ httpie/input.py
@@ -171,12 +171,10 @@ class HTTPieArgumentParser(ArgumentParser):
         Modify `env.stdout` and `env.stdout_isatty` based on args, if needed.
 
         """
-        if not self.env.stdout_isatty and self.args.output_file:
-            self.error('Cannot use --output, -o with redirected output.')
-
+        self.args.output_file_specified = bool(self.args.output_file)
         if self.args.download:
             # FIXME: Come up with a cleaner solution.
-            if not self.env.stdout_isatty:
+            if not self.args.output_file and not self.env.stdout_isatty:
                 # Use stdout as the download output file.
                 self.args.output_file = self.env.stdout
             # With `--download`, we write everything that would normally go to